gen-api/pyproject.toml
[project]
name = "gen-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "boto3~=1.26.136",
    "celery==5.3.6",
    "django==5.0.7",
    "django-cors-headers==3.13.0",
    "djangorestframework==3.15.2",
    "drf-yasg==1.21.5",
    "gunicorn==22.0.0",
    "markdown==3.4.4",
    "mock~=5.0.1",
    "pyjwt==2.8.0",
    "setuptools~=69.1.1",
]
